+++ src/command-parser.ts
@@ -1,6 +1,5 @@
 import { Command, Commands } from "./definitions";
 
-
 /**
  * Return the Command object the command string specifies.
  * Returns false if the command string did not have a valid command;
@@ -9,18 +8,18 @@ import { Command, Commands } from "./definitions";
 export function parseCommand(promptResponse: string[], commands: Commands): (
     { command: Command, remainingPieces: string[] } | false
 ) {
-   const command = commands[promptResponse[0]];
+    const command = commands[promptResponse[0]];
 
-   if (!command) return false;
+    if (!command) return false;
 
-   if (command.subcommands) {
-       const subcommand = parseCommand(promptResponse.slice(1), command.subcommands);
-       if (subcommand) {
-           return subcommand;
-       }
-   }
+    if (command.subcommands) {
+        const subcommand = parseCommand(promptResponse.slice(1), command.subcommands);
+        if (subcommand) {
+            return subcommand;
+        }
+    }
 
-   return {
-       command, remainingPieces: promptResponse.slice(1)
-   }
+    return {
+        command, remainingPieces: promptResponse.slice(1),
+    };
 }