+++ src/example.ts
@@ -1,52 +1,46 @@
 import { CLI } from "./index";
+const { log } = console;
 
-const cli = new CLI();
-
-cli.setDelimiter("example -> ")
-
-cli.command("hello", {
-    action: (params, opts, done) => {
-        console.log("Hello Back!");
-        done();
-    }
-});
-
-cli.command("hide", {
-    action() {
-        cli.hide();
-        setTimeout(() => {
-            cli.show();
-        }, 2000);
-        return Promise.resolve();
-    }
-});
-
-cli.command("do", {
-    description: "Preform an action",
-    action: (a, b, c) => c(),
-    subcommands: {
-        something: {
-            options: [
-                "joint",
-                "hi"
-            ],
-            parameters: [
-                { label: "num", type: "number" }
-            ],
-            action: (a, b, c) => {
-                console.log(`I Did Something!`);
-                console.log(a);
-                console.log(b);
-                c();
-            }
+const cli = new CLI()
+    .setDelimiter("example -> ")
+    .addExitCommand()
+    .addCommand("hello", {
+        action(parameters, options) {
+            log("Hello Back!");
         },
-        nothing: {
-            action: (a, b, c) => {
-                console.log(`I Did Nothing!`);
-                c();
-            }
-        }
-    }
-});
-
-cli.show();
+    })
+    .addCommand("hide", {
+        action(parameters, options) {
+            cli.hide();
+            setTimeout(() => {
+                cli.show();
+            }, 2000);
+            return Promise.resolve();
+        },
+    })
+    .addCommand("do", {
+        description: "Preform an action",
+        action: (parameters, options) => null,
+        subcommands: {
+            something: {
+                options: [
+                    "joint",
+                    "hi",
+                ],
+                parameters: [
+                    { label: "num", type: "number" },
+                ],
+                action(parameters, options) {
+                    log("I Did Something!");
+                    log(parameters.rer);
+                    log(options);
+                },
+            },
+            nothing: {
+                action(parameters, options) {
+                    log("I Did Nothing!");
+                },
+            },
+        },
+    })
+    .show();