+++ src/help-gen.ts
@@ -1,11 +1,12 @@
 import { Command } from "./definitions";
 const indent = require("indent");
 const columnify = require("columnify");
+const { log } = console;
 
 function columns(data: any) {
     return indent(columnify(data, {
         showHeaders: false,
-        minWidth: 30
+        minWidth: 30,
     }), 4);
 }
 
@@ -13,60 +14,64 @@ function genUsage(command: string, val: Command) {
     let usage = `${command} [options]`;
     if (!val.parameters) return usage;
     for (const param of val.parameters) {
-        usage += ` <${param.label}>`
+        usage += ` <${param.label}>`;
     }
     return usage;
 }
 
 export function printCommandHelp(command: string, val: Command) {
-    console.log("");
+    log("");
     if (val.description) {
-        console.log(val.description + "\n");
+        log(val.description + "\n");
     }
-    console.log(`Usage:\n`);
-    console.log(indent(genUsage(command, val), 4) + "\n");
+    log("Usage:\n");
+    log(indent(genUsage(command, val), 4) + "\n");
     printOptions(val);
     printSubCommands(val);
-    console.log("");
+    log("");
 }
 
 function printOptions(val: Command) {
     if (val.options && val.options.length > 0) {
-        console.log(`Options:\n`);
+        log("Options:\n");
         const options: any = {};
-        val.options.forEach(opt => {
+        val.options.forEach((opt) => {
             if (typeof opt === "string") {
                 options[`@${opt}`] = "";
                 return;
             }
             options[`@${opt}`] = opt.description || "";
         });
-        console.log(columns(options));
+        log(columns(options));
     }
 }
 
 function printSubCommands(val: Command) {
     if (val.subcommands) {
-        console.log("Sub-Commands:\n");
+        log("Sub-Commands:\n");
         const commands: any = {};
         for (const command in val.subcommands) {
-            commands[genUsage(command, val.subcommands[command])] = val.subcommands[command].description || "";
+            if (val.subcommands.hasOwnProperty(command)) {
+                commands[genUsage(command, val.subcommands[command])] = val.subcommands[command].description || "";
+            }
         }
-        console.log(columns(commands));
+        log(columns(commands));
     }
 }
 
 export function printOverviewHelp(commands: { [command: string]: Command }) {
     const commandDescriptions: any = {};
     for (const command in commands) {
-        commandDescriptions[genUsage(command, commands[command])] = commands[command].description || "";
+        if (commands.hasOwnProperty(command)) {
+            commandDescriptions[genUsage(command, commands[command])] = commands[command].description || "";
+        }
     }
-    console.log(`\nUsage:\n`)
-    console.log(indent(`<cmd>\n`, 4));
+    log("\nUsage:\n");
+    log(indent("<cmd>\n", 4));
 
-    console.log(`Available commands:\n`);
-    console.log(indent(`help <cmd>`, 4));
+    log("Available commands:\n");
+    log(indent("help <cmd>", 4));
 
-    console.log(columns(commandDescriptions) + "\n");
+    log(columns(commandDescriptions) + "\n");
     return;
 }